moved the coders into separate files
